gqlgenのTracerの実装をやり直したので実験する
